Move prompts in context directory
- Move prompt.py from context/utils/ to context/prompt.py
- Move template files from context/utils/prompts/ to context/prompts/templates/
- Update import statements in all affected files
- Remove old utils directory structure
- All tests pass after restructuring

Fixes #190
