+++ openhands/sdk/context/prompts/__init__.py
@@ -1,0 +1,1 @@
+# This module contains prompt templates and utilities
