+++ openhands/sdk/context/prompts/templates/microagent_knowledge_info.j2
