+++ openhands/sdk/context/prompts/templates/system_message_suffix.j2
