--- openhands/sdk/context/utils/__init__.py
@@ -1,8 +1,0 @@
-from openhands.sdk.context.utils.prompt import (
-    render_template,
-)
-
-
-__all__ = [
-    "render_template",
-]
